terms view to its own file, paste terms text
